Addition of column splitter for parallel lang text
This adds a feature to allow you to take content from the nth column of
all table rows and extract only that content back into the document.
Content in other columns is removed. Good if you have multilingual docs
with content in side-by-side columns, and want to make the doc
unilingual.